+++ tests/io/test_readers_sequences_json.py
@@ -3,6 +3,7 @@ import json
 
 import pytest
 
+from datarec import DataRec
 from datarec.io.readers.sequences.json import (
     read_sequences_json,
     read_sequences_json_array,
@@ -31,9 +32,10 @@ def test_read_sequences_json(tmp_path):
         rating_col="rating",
         timestamp_col="timestamp",
     )
+    assert isinstance(dr, DataRec)
     assert len(dr.data) == 3
-    assert set(dr.data[dr._user_col]) == {"u1", "u2"}
-    assert set(dr.data[dr._item_col]) == {"i1", "i2", "i3"}
+    assert set(dr.data[dr.user_col]) == {"u1", "u2"}
+    assert set(dr.data[dr.item_col]) == {"i1", "i2", "i3"}
 
 
 def test_read_sequences_json_array(tmp_path):
@@ -52,9 +54,10 @@ def test_read_sequences_json_array(tmp_path):
         sequence_key="sequence",
         rating_col="rating",
     )
+    assert isinstance(rd, DataRec)
     assert len(rd.data) == 3
-    assert set(rd.data[rd._user_col]) == {"u1", "u2"}
-    assert set(rd.data[rd._item_col]) == {1, 2, 3}
+    assert set(rd.data[rd.user_col]) == {"u1", "u2"}
+    assert set(rd.data[rd.item_col]) == {1, 2, 3}
 
 
 def test_read_sequences_json_items(tmp_path):
@@ -67,11 +70,12 @@ def test_read_sequences_json_items(tmp_path):
         user_col="user",
         item_col="item",
     )
+    assert isinstance(rd, DataRec)
 
     assert len(rd.data) == 4
-    assert set(rd.data[rd._user_col]) == {"0", "1"}
-    assert list(rd.data[rd.data[rd._user_col] == "0"][rd._item_col]) == [4755, 9066, 9765]
-    assert list(rd.data[rd.data[rd._user_col] == "1"][rd._item_col]) == [2975]
+    assert set(rd.data[rd.user_col]) == {"0", "1"}
+    assert list(rd.data[rd.data[rd.user_col] == "0"][rd.item_col]) == [4755, 9066, 9765]
+    assert list(rd.data[rd.data[rd.user_col] == "1"][rd.item_col]) == [2975]
 
 
 def test_read_sequences_json_items_requires_object(tmp_path):